unused files removed and const type added to enums
