+++ src/index.test.ts
@@ -1,4 +1,4 @@
-import HighlightText from "./index"
+import HighlightText, { QuotationType } from "./index"
 
 test("Basic typescript search", () => {
   const text = new HighlightText()
@@ -7,4 +7,22 @@ test("Basic typescript search", () => {
   expect(replaceWord).toEqual(
     "<span class='highlight-text'>Merhaba</span> Dünya"
   )
+})
+
+test("Tag should be same as given", () => {
+  const highlight = new HighlightText({
+    tag: "i"
+  })
+  const result = highlight.replaceWord("Merhaba Dünya", "Dünya")
+
+  expect(result).toEqual("Merhaba <i class='highlight-text'>Dünya</i>")
+})
+
+test("Quote should be double", () => {
+  const highlight = new HighlightText({
+    quotation: QuotationType.Single
+  })
+  const result = highlight.replaceWord("Merhaba Dünya", "Dünya")
+
+  expect(result).toEqual("Merhaba <span class='highlight-text'>Dünya</span>")
 })
